requirements.txt
bcrypt==4.0.1
cffi==1.15.1
cryptography==39.0.0
Fabric3==1.14.post1
paramiko==2.12.0
pycparser==2.21
PyNaCl==1.5.0
python-dotenv==0.21.0
python-vagrant==1.0.0
six==1.16.0
